+++ src/django_typer/apps.py
@@ -30,28 +30,25 @@ def check_traceback_config(app_configs, **kwargs) -> t.List[CheckMessage]:
     warnings: t.List[CheckMessage] = []
     if use_rich_tracebacks():
         tb_config = traceback_config()
-        try:
-            from rich import traceback  # pyright: ignore[reportMissingImports]
+        from rich import traceback  # pyright: ignore[reportMissingImports]
 
-            expected = {
-                "no_install",
-                "short",
-                *inspect.signature(traceback.install).parameters.keys(),
-            }
-            unexpected = set(tb_config.keys()) - expected
-            if unexpected:
-                warnings.append(
-                    CheckWarning(
-                        "DT_RICH_TRACEBACK_CONFIG",
-                        hint="Unexpected parameters encountered: {keys}.".format(
-                            keys=", ".join(unexpected)
-                        ),
-                        obj=settings.SETTINGS_MODULE,
-                        id="django_typer.W001",
-                    )
+        expected = {
+            "no_install",
+            "short",
+            *inspect.signature(traceback.install).parameters.keys(),
+        }
+        unexpected = set(tb_config.keys()) - expected
+        if unexpected:
+            warnings.append(
+                CheckWarning(
+                    "DT_RICH_TRACEBACK_CONFIG",
+                    hint="Unexpected parameters encountered: {keys}.".format(
+                        keys=", ".join(unexpected)
+                    ),
+                    obj=settings.SETTINGS_MODULE,
+                    id="django_typer.W001",
                 )
-        except ImportError:
-            pass
+            )
     return warnings