+++ tests/shellcompletion/test_powershell.py
@@ -58,9 +58,12 @@ class _PowerShellMixin:
     def test_mixed_path_dividers(self):
         self.install()
         self.verify_install()
+        completions = self.get_completions(
+            "completion", "--path", "./src/django_typer\\comp"
+        )
         self.assertIn(
-            "django_typer\\completers",
-            self.get_completions("completion", "--path", "./django_typer\\comp"),
+            "src\\django_typer\\completers",
+            completions,
         )
         self.remove()
         self.verify_remove()
@@ -156,8 +159,8 @@ class PWSHExeTests(_PowerShellMixin, _InstalledScriptCompleteTestCase, TestCase)
         self.install()
         self.verify_install()
         self.assertIn(
-            "django_typer\\completers",
-            self.get_completions("completion", "--path", "./django_typer\\comp"),
+            "src\\django_typer\\completers",
+            self.get_completions("completion", "--path", "./src/django_typer\\comp"),
         )
         self.remove()
         self.verify_remove()