+++ doc/source/changelog.rst
@@ -4,6 +4,17 @@
 Change Log
 ==========
 
+v3.1.0 (2024-03-xx)
+===================
+
+* Fixed `Shell completion tests let failures through in CI <https://github.com/django-commons/django-typer/issues/194>`_
+* Fixed `fish completion installs should respect XDG_CONFIG_HOME <https://github.com/django-commons/django-typer/issues/193>`_
+* Fixed `zsh completion installs should respect ZDOTDIR <https://github.com/django-commons/django-typer/issues/192>`_
+* Implemented `Support Django 5.2 <https://github.com/django-commons/django-typer/issues/188>`_
+* Implemented `Switch poetry -> uv <https://github.com/django-commons/django-typer/issues/185>`_
+* Implemented `Require tests to pass before release action runs. <<https://github.com/django-commons/django-typer/issues/173>`_
+
+
 v3.0.0 (2024-02-16)
 ===================